client: implement webxr renderstate updates

+++ src/xr/render_state.hpp
@@ -52,6 +52,6 @@ namespace xr
     WebGLLayer *baseLayer = nullptr;
     float depthFar;
     float depthNear;
-    uint32_t inlineVerticalFieldOfView;
+    float inlineVerticalFieldOfView;
   };
 }
